+++ src/sigmaplus_verifier.hpp
@@ -51,7 +51,7 @@ bool SigmaPlusVerifier<Exponent, GroupElement>::verify(
     f_i_.reserve(N);
     for(int i = 0; i < N; ++i) {
         std::vector<uint64_t> I = SigmaPrimitives<Exponent, GroupElement>::convert_to_nal(i, n, m);
-        Exponent f_i(uint64_t(1));
+        Exponent f_i(unsigned(1));
         for(int j = 0; j < m; ++j){
             f_i *= f[j*n + I[j]];
         }
@@ -61,14 +61,14 @@ bool SigmaPlusVerifier<Exponent, GroupElement>::verify(
     secp_primitives::MultiExponent mult(commits, f_i_);
     GroupElement t1 = mult.get_multiple();
     GroupElement t2;
-    Exponent x_k(uint64_t(1));
+    Exponent x_k(unsigned(1));
     for(int k = 0; k < m; ++k){
         t2 += (Gk[k] * (x_k.negate()));
         x_k *= challenge_x;
     }
 
     GroupElement left(t1 + t2);
-    if (left != SigmaPrimitives<Exponent, GroupElement>::commit(g_, Exponent(uint64_t(0)), h_[0], proof.z_))
+    if (left != SigmaPrimitives<Exponent, GroupElement>::commit(g_, Exponent(unsigned(0)), h_[0], proof.z_))
         return false;
 
     return true;
